Add experimental blop implementation
Not sure how I feel about this. The reason why making it be exactly like the figma design is not viable, is that we do the thing where we center the content if the screen is too big, like the LNU website. Making the blob fixed to the left edge doesn't make much sense in that case